+++ builds/spacetime.cjs
@@ -1,4 +1,4 @@
-/* spencermountain/spacetime 7.14.0 Apache 2.0 */
+/* spencermountain/spacetime 7.15.0 Apache 2.0 */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -23,6 +23,13 @@
     return epoch >= startUtc && epoch < endUtc
   };
 
+  // Keep metadata and the fast wall-clock calculation in agreement.
+  var getDstShift = tz => {
+    if (tz === 'australia/lord_howe') return 0.5
+    if (tz === 'antarctica/troll') return 2
+    return 1
+  };
+
   /* eslint-disable no-console */
 
   // this method avoids having to do a full dst-calculation on every operation
@@ -40,9 +47,10 @@
 
     //get our two possible offsets
     const jul = obj.offset;
-    let dec = obj.offset + 1; // assume it's the same for now
+    const shift = getDstShift(s.tz);
+    let dec = obj.offset + shift;
     if (obj.hem === 'n') {
-      dec = jul - 1;
+      dec = jul - shift;
     }
     const split = obj.dst.split('->');
     const inSummer = inSummerTime(s.epoch, split[0], split[1], jul, dec);
@@ -57,21 +65,19 @@
     "9|n": "2/chita,2/khandyga,2/pyongyang,2/seoul,2/tokyo,2/yakutsk,11/palau,japan,rok",
     "9.5|s|04/05:03->10/04:02": "4/adelaide,4/broken_hill,4/south,4/yancowinna",
     "9.5|s": "4/darwin,4/north",
-    "8|s|03/13:01->10/02:00": "12/casey",
-    "8|s": "2/kuala_lumpur,2/makassar,2/singapore,4/perth,2/ujung_pandang,4/west,singapore",
+    "8|s": "12/casey,2/kuala_lumpur,2/makassar,2/singapore,4/perth,2/ujung_pandang,4/west,singapore",
     "8|n": "2/brunei,2/hong_kong,2/irkutsk,2/kuching,2/macau,2/manila,2/shanghai,2/taipei,2/ulaanbaatar,2/chongqing,2/chungking,2/harbin,2/macao,2/ulan_bator,2/choibalsan,hongkong,prc,roc",
     "8.75|s": "4/eucla",
     "7|s": "12/davis,2/jakarta,9/christmas",
     "7|n": "2/bangkok,2/barnaul,2/hovd,2/krasnoyarsk,2/novokuznetsk,2/novosibirsk,2/phnom_penh,2/pontianak,2/ho_chi_minh,2/tomsk,2/vientiane,2/saigon",
-    "6|s": "12/vostok",
     "6|n": "2/bishkek,2/dhaka,2/omsk,2/thimphu,2/urumqi,9/chagos,2/dacca,2/kashgar,2/thimbu",
     "6.5|n": "2/yangon,9/cocos,2/rangoon",
-    "5|s": "12/mawson,9/kerguelen",
+    "5|s": "12/mawson,12/vostok,9/kerguelen",
     "5|n": "2/almaty,2/aqtau,2/aqtobe,2/ashgabat,2/atyrau,2/dushanbe,2/karachi,2/oral,2/qyzylorda,2/qostanay,2/samarkand,2/tashkent,2/yekaterinburg,9/maldives,2/ashkhabad",
     "5.75|n": "2/kathmandu,2/katmandu",
     "5.5|n": "2/kolkata,2/colombo,2/calcutta",
     "4|s": "9/reunion",
-    "4|n": "2/baku,2/dubai,2/muscat,2/tbilisi,2/yerevan,8/astrakhan,8/samara,8/saratov,8/ulyanovsk,8/volgograd,9/mahe,9/mauritius,2/volgograd",
+    "4|n": "2/baku,2/dubai,2/muscat,2/tbilisi,2/yerevan,8/astrakhan,8/samara,8/saratov,8/ulyanovsk,9/mahe,9/mauritius",
     "4.5|n": "2/kabul",
     "3|s": "12/syowa,9/antananarivo",
     "3|n|04/24:00->10/29:24": "0/cairo,egypt",
@@ -79,13 +85,13 @@
     "3|n|03/29:00->10/24:24": "2/beirut",
     "3|n|03/28:02->10/24:02": "2/gaza,2/hebron",
     "3|n|03/27:02->10/25:02": "2/jerusalem,2/tel_aviv,israel",
-    "3|n": "0/addis_ababa,0/asmara,0/asmera,0/dar_es_salaam,0/djibouti,0/juba,0/kampala,0/mogadishu,0/nairobi,2/aden,2/amman,2/baghdad,2/bahrain,2/damascus,2/kuwait,2/qatar,2/riyadh,8/istanbul,8/kirov,8/minsk,8/moscow,8/simferopol,9/comoro,9/mayotte,2/istanbul,turkey,w-su",
+    "3|n": "0/addis_ababa,0/asmara,0/asmera,0/dar_es_salaam,0/djibouti,0/kampala,0/mogadishu,0/nairobi,2/aden,2/amman,2/baghdad,2/bahrain,2/damascus,2/kuwait,2/qatar,2/riyadh,8/istanbul,8/kirov,8/minsk,8/moscow,8/simferopol,8/volgograd,9/comoro,9/mayotte,2/volgograd,2/istanbul,turkey,w-su",
     "3.5|n": "2/tehran,iran",
-    "2|s|03/29:02->10/25:02": "12/troll",
-    "2|s": "0/gaborone,0/harare,0/johannesburg,0/lubumbashi,0/lusaka,0/maputo,0/maseru,0/mbabane",
+    "2|s": "0/gaborone,0/harare,0/johannesburg,0/lubumbashi,0/lusaka,0/maputo,0/maseru,0/mbabane,0/windhoek",
     "2|n|03/29:02->10/25:03": "0/ceuta,arctic/longyearbyen,8/amsterdam,8/andorra,8/belgrade,8/berlin,8/bratislava,8/brussels,8/budapest,8/busingen,8/copenhagen,8/gibraltar,8/ljubljana,8/luxembourg,8/madrid,8/malta,8/monaco,8/oslo,8/paris,8/podgorica,8/prague,8/rome,8/san_marino,8/sarajevo,8/skopje,8/stockholm,8/tirane,8/vaduz,8/vatican,8/vienna,8/warsaw,8/zagreb,8/zurich,3/jan_mayen,poland,cet,met",
-    "2|n": "0/blantyre,0/bujumbura,0/khartoum,0/kigali,0/tripoli,8/kaliningrad,libya",
-    "1|s": "0/brazzaville,0/kinshasa,0/luanda,0/windhoek",
+    "2|n|03/29:01->10/25:03": "12/troll",
+    "2|n": "0/blantyre,0/bujumbura,0/juba,0/khartoum,0/kigali,0/tripoli,8/kaliningrad,libya",
+    "1|s": "0/brazzaville,0/kinshasa,0/luanda",
     "1|n|03/29:01->10/25:02": "3/canary,3/faroe,3/madeira,8/dublin,8/guernsey,8/isle_of_man,8/jersey,8/lisbon,8/london,3/faeroe,eire,8/belfast,gb-eire,gb,portugal,wet",
     "1|n": "0/algiers,0/bangui,0/douala,0/lagos,0/libreville,0/malabo,0/ndjamena,0/niamey,0/porto-novo,0/tunis",
     "14|n": "11/kiritimati",
@@ -93,49 +99,46 @@
     "13|n": "11/enderbury,11/kanton,11/fakaofo",
     "12|s|04/05:03->09/27:02": "12/mcmurdo,11/auckland,12/south_pole,nz",
     "12|s": "11/fiji",
-    "12|n": "2/anadyr,2/kamchatka,2/srednekolymsk,11/funafuti,11/kwajalein,11/majuro,11/nauru,11/tarawa,11/wake,11/wallis,kwajalein",
-    "12.75|s|04/05:03->04/05:02": "11/chatham,nz-chat",
-    "11|s|04/05:03->10/04:02": "12/macquarie",
+    "12|n": "2/anadyr,2/kamchatka,11/funafuti,11/kwajalein,11/majuro,11/nauru,11/tarawa,11/wake,11/wallis,kwajalein",
+    "12.75|s|04/05:03->09/27:02": "11/chatham,nz-chat",
+    "11|s|04/05:03->10/04:02": "11/norfolk",
     "11|s": "11/bougainville",
-    "11|n": "2/magadan,2/sakhalin,11/efate,11/guadalcanal,11/kosrae,11/noumea,11/pohnpei,11/ponape",
-    "11.5|n|04/05:03->10/04:02": "11/norfolk",
-    "10|s|04/05:03->10/04:02": "4/currie,4/hobart,4/melbourne,4/sydney,4/act,4/canberra,4/nsw,4/tasmania,4/victoria",
+    "11|n": "2/magadan,2/sakhalin,2/srednekolymsk,11/efate,11/guadalcanal,11/kosrae,11/noumea,11/pohnpei,11/ponape",
+    "10|s|04/05:03->10/04:02": "12/macquarie,4/currie,4/hobart,4/melbourne,4/sydney,4/act,4/canberra,4/nsw,4/tasmania,4/victoria",
     "10|s": "12/dumontdurville,4/brisbane,4/lindeman,11/port_moresby,4/queensland",
     "10|n": "2/ust-nera,2/vladivostok,11/guam,11/saipan,11/chuuk,11/truk,11/yap",
-    "10.5|s|04/05:01->10/04:02": "4/lord_howe,4/lhi",
-    "0|s|02/15:03->03/22:02": "0/casablanca,0/el_aaiun",
+    "10.5|s|04/05:02->10/04:02": "4/lord_howe,4/lhi",
+    "0|s|02/15:03->09/20:02": "0/casablanca",
+    "0|s|02/15:03->03/22:02": "0/el_aaiun",
     "0|n|03/29:00->10/25:01": "3/azores",
-    "0|n|03/29:00->10/24:24": "1/scoresbysund",
     "0|n": "0/abidjan,0/accra,0/bamako,0/banjul,0/bissau,0/conakry,0/dakar,0/freetown,0/lome,0/monrovia,0/nouakchott,0/ouagadougou,0/sao_tome,1/danmarkshavn,3/reykjavik,3/st_helena,13/gmt,13/utc,0/timbuktu,13/greenwich,13/uct,13/universal,13/zulu,gmt-0,gmt+0,gmt0,greenwich,iceland,uct,universal,utc,zulu,13/unknown,factory",
     "-9|n|03/08:02->11/01:02": "1/adak,1/atka,us/aleutian",
     "-9|n": "11/gambier",
     "-9.5|n": "11/marquesas",
     "-8|n|03/08:02->11/01:02": "1/anchorage,1/juneau,1/metlakatla,1/nome,1/sitka,1/yakutat,us/alaska",
     "-8|n": "11/pitcairn",
-    "-7|n|03/08:02->11/01:03": "1/vancouver,6/pacific",
+    "-7|n|03/08:02->12/31:23": "1/vancouver,6/pacific",
     "-7|n|03/08:02->11/01:02": "1/los_angeles,1/santa_isabel,1/tijuana,1/ensenada,10/bajanorte,us/pacific-new,us/pacific",
     "-7|n": "1/creston,1/dawson,1/dawson_creek,1/fort_nelson,1/hermosillo,1/mazatlan,1/phoenix,1/whitehorse,6/yukon,10/bajasur,us/arizona,mst",
     "-6|s|04/04:22->09/05:22": "11/easter,7/easterisland",
-    "-6|n|04/07:02->10/27:02": "1/merida",
-    "-6|n|03/08:02->11/01:03": "1/edmonton,1/yellowknife,6/mountain",
-    "-6|n|03/08:02->11/01:02": "1/boise,1/cambridge_bay,1/denver,1/inuvik,1/north_dakota,1/ojinaga,1/ciudad_juarez,1/shiprock,navajo,us/mountain",
-    "-6|n": "1/bahia_banderas,1/belize,1/chihuahua,1/costa_rica,1/el_salvador,1/guatemala,1/managua,1/mexico_city,1/monterrey,1/regina,1/swift_current,1/tegucigalpa,11/galapagos,6/east-saskatchewan,6/saskatchewan,10/general",
+    "-6|n|03/08:02->12/31:23": "1/edmonton,1/yellowknife,6/mountain",
+    "-6|n|03/08:02->11/01:02": "1/boise,1/cambridge_bay,1/denver,1/inuvik,1/north_dakota,1/ciudad_juarez,1/shiprock,navajo,us/mountain",
+    "-6|n": "1/bahia_banderas,1/belize,1/chihuahua,1/costa_rica,1/el_salvador,1/guatemala,1/managua,1/merida,1/mexico_city,1/monterrey,1/regina,1/swift_current,1/tegucigalpa,11/galapagos,6/east-saskatchewan,6/saskatchewan,10/general",
     "-5|s": "1/lima,1/rio_branco,1/porto_acre,5/acre",
-    "-5|n|03/08:02->11/01:02": "1/chicago,1/matamoros,1/menominee,1/rainy_river,1/rankin_inlet,1/resolute,1/winnipeg,1/indiana/knox,1/indiana/tell_city,1/north_dakota/beulah,1/north_dakota/center,1/north_dakota/new_salem,1/knox_in,6/central,us/central,us/indiana-starke",
+    "-5|n|03/08:02->11/01:02": "1/chicago,1/matamoros,1/menominee,1/ojinaga,1/rainy_river,1/rankin_inlet,1/resolute,1/winnipeg,1/indiana/knox,1/indiana/tell_city,1/north_dakota/beulah,1/north_dakota/center,1/north_dakota/new_salem,1/knox_in,6/central,us/central,us/indiana-starke",
     "-5|n": "1/bogota,1/cancun,1/cayman,1/coral_harbour,1/eirunepe,1/guayaquil,1/jamaica,1/panama,1/atikokan,jamaica,est",
     "-4|s|04/04:24->09/06:00": "1/santiago,7/continental",
-    "-4|s|03/22:24->10/05:00": "1/asuncion",
     "-4|s": "1/campo_grande,1/cuiaba,1/la_paz,1/manaus,5/west",
     "-4|n|03/08:02->11/01:02": "1/detroit,1/grand_turk,1/indiana,1/indianapolis,1/iqaluit,1/kentucky,1/louisville,1/montreal,1/nassau,1/new_york,1/nipigon,1/pangnirtung,1/port-au-prince,1/thunder_bay,1/toronto,1/indiana/marengo,1/indiana/petersburg,1/indiana/vevay,1/indiana/vincennes,1/indiana/winamac,1/kentucky/monticello,1/fort_wayne,1/indiana/indianapolis,1/kentucky/louisville,6/eastern,us/east-indiana,us/eastern,us/michigan",
     "-4|n|03/08:00->11/01:01": "1/havana,cuba",
     "-4|n": "1/anguilla,1/antigua,1/aruba,1/barbados,1/blanc-sablon,1/boa_vista,1/caracas,1/curacao,1/dominica,1/grenada,1/guadeloupe,1/guyana,1/kralendijk,1/lower_princes,1/marigot,1/martinique,1/montserrat,1/port_of_spain,1/porto_velho,1/puerto_rico,1/santo_domingo,1/st_barthelemy,1/st_kitts,1/st_lucia,1/st_thomas,1/st_vincent,1/tortola,1/virgin",
-    "-3|s": "1/argentina,1/buenos_aires,1/catamarca,1/cordoba,1/coyhaique,1/fortaleza,1/jujuy,1/mendoza,1/montevideo,1/punta_arenas,1/sao_paulo,12/palmer,12/rothera,3/stanley,1/argentina/la_rioja,1/argentina/rio_gallegos,1/argentina/salta,1/argentina/san_juan,1/argentina/san_luis,1/argentina/tucuman,1/argentina/ushuaia,1/argentina/comodrivadavia,1/argentina/buenos_aires,1/argentina/catamarca,1/argentina/cordoba,1/argentina/jujuy,1/argentina/mendoza,1/argentina/rosario,1/rosario,5/east",
+    "-3|s": "1/argentina,1/asuncion,1/buenos_aires,1/catamarca,1/cordoba,1/coyhaique,1/fortaleza,1/jujuy,1/mendoza,1/montevideo,1/punta_arenas,1/sao_paulo,12/palmer,12/rothera,3/stanley,1/argentina/la_rioja,1/argentina/rio_gallegos,1/argentina/salta,1/argentina/san_juan,1/argentina/san_luis,1/argentina/tucuman,1/argentina/ushuaia,1/argentina/comodrivadavia,1/argentina/buenos_aires,1/argentina/catamarca,1/argentina/cordoba,1/argentina/jujuy,1/argentina/mendoza,1/argentina/rosario,1/rosario,5/east",
     "-3|n|03/08:02->11/01:02": "1/glace_bay,1/goose_bay,1/halifax,1/moncton,1/thule,3/bermuda,6/atlantic",
     "-3|n": "1/araguaina,1/bahia,1/belem,1/cayenne,1/maceio,1/paramaribo,1/recife,1/santarem",
     "-2|n|03/08:02->11/01:02": "1/miquelon",
     "-2|n": "1/noronha,3/south_georgia,5/denoronha",
     "-2.5|n|03/08:02->11/01:02": "1/st_johns,6/newfoundland",
-    "-1|n|03/29:00->10/24:24": "1/nuuk,1/godthab",
+    "-1|n|03/28:23->10/24:24": "1/nuuk,1/scoresbysund,1/godthab",
     "-1|n": "3/cape_verde",
     "-11|n": "11/midway,11/niue,11/pago_pago,11/samoa,us/samoa",
     "-10|n": "11/honolulu,11/johnston,11/rarotonga,11/tahiti,us/hawaii,hst"
@@ -1101,8 +1104,8 @@
           s.epoch = null;
           return s
         }
-        walkTo(s, obj);
         s = parseOffset(s, tz);
+        walkTo(s, obj);
         s = parseTime(s, time);
         return s
       }
@@ -1122,8 +1125,8 @@
           s.epoch = null;
           return s
         }
-        walkTo(s, obj);
         s = parseOffset(s, tz);
+        walkTo(s, obj);
         s = parseTime(s, time);
         return s
       }
@@ -2595,7 +2598,8 @@
     const summer = found.offset; // (july)
     let winter = summer; // (january) assume it's the same for now
     //most zones shift by 1hr on dst, but Lord Howe is a ½-hour shift
-    const dstShift = tz === 'australia/lord_howe' ? 0.5 : 1;
+    //and Troll (Antarctica) makes a 2-hour jump between +00 and +02
+    const dstShift = getDstShift(tz);
     if (result.hasDst === true) {
       if (result.hemisphere === 'North') {
         winter = summer - dstShift;
@@ -4200,7 +4204,7 @@
     return tzs
   };
 
-  var version = '7.14.0';
+  var version = '7.15.0';
 
   const main = (input, tz, options) => new SpaceTime(input, tz, options);