+++ src/timezone/index.js
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 import findTz from './find.js'
 import inSummerTime from './summerTime.js'
+import getDstShift from './dstShift.js'
 
 const parseDst = dst => {
   if (!dst) {
@@ -66,7 +67,8 @@ const timezone = s => {
   const summer = found.offset // (july)
   let winter = summer // (january) assume it's the same for now
   //most zones shift by 1hr on dst, but Lord Howe is a ½-hour shift
-  const dstShift = tz === 'australia/lord_howe' ? 0.5 : 1
+  //and Troll (Antarctica) makes a 2-hour jump between +00 and +02
+  const dstShift = getDstShift(tz)
   if (result.hasDst === true) {
     if (result.hemisphere === 'North') {
       winter = summer - dstShift